requirements.txt
backtrader
pandas
matplotlib
jupyter
TA-Lib
bta-lib
ta
requests
python-dotenv
tabulate
seaborn
mplfinance
plotly
dash
bayesian-optimization
optuna
kiteconnect
streamlit
pymongo[srv]